refactor: ebird_region_core shared module; processor drops services.* (#238)
- New app/ebird_region_core.py: region code, eBird HTTP top-N, TTL cache, name mapping
- ebird_region_service re-exports from core; get_region_comparison unchanged (web cache)
- ebird_util.REGION_NAME_TO_CODE from core; Dockerfile COPY ebird_region_core.py
- Tests: patch ebird_region_core; processor test adds /app root to path

Made-with: Cursor
