+++ app/web/services/ebird_region_service.py
@@ -1,107 +1,23 @@
-"""eBird API: regional species for comparison with user's feeder."""
+"""eBird API: regional species for comparison with user's feeder (web cache layer)."""
 import hashlib
 import logging
-import time
-from collections import Counter
-
-import httpx
 
 from app_config.app_config import app_config
 
+from ebird_region_core import (
+    _REGION_TOP_CACHE,
+    _build_region_code,
+    ebird_common_to_birdlense_name,
+    get_region_top_species,
+    get_region_top_species_cached,
+)
 from services.cache import cache_get, cache_set
-from services.ebird_util import REGION_NAME_TO_CODE, common_name_from_species
+from services.ebird_util import common_name_from_species
 
 logger = logging.getLogger(__name__)
 
-EBIRD_API_BASE = 'https://api.ebird.org/v2'
-TOP_N = 20
-BACK_DAYS = 30
-MAX_OBSERVATIONS = 5000
-
-
-def _build_region_code() -> str:
-    """Build eBird region code from config: country or country-state."""
-    country = (app_config.get("ebird.country") or "").strip().upper() or "US"
-    state_raw = (app_config.get("ebird.state") or "").strip()
-    if not state_raw:
-        return country
-    # Поддержка полных названий: "Moscow Oblast" -> MOS
-    state_lower = state_raw.lower().strip()
-    state = REGION_NAME_TO_CODE.get(state_lower) or state_raw.upper()[:3]
-    return f"{country}-{state}"
-
-
-def _fetch_region_observations(api_key: str, region_code: str) -> list[dict]:
-    """Fetch recent observations from eBird API. Returns list of obs dicts."""
-    url = f'{EBIRD_API_BASE}/data/obs/{region_code}/recent'
-    params = {'back': BACK_DAYS, 'maxResults': MAX_OBSERVATIONS}
-    headers = {'X-eBirdApiToken': api_key}
-    try:
-        with httpx.Client(timeout=15.0) as client:
-            r = client.get(url, params=params, headers=headers)
-            r.raise_for_status()
-            return r.json() or []
-    except httpx.HTTPStatusError as e:
-        txt = e.response.text[:200]
-        logger.warning('eBird API error %s: %s', e.response.status_code, txt)
-        raise ValueError(f'eBird API: {e.response.status_code}') from e
-    except Exception as e:
-        logger.warning('eBird fetch failed: %s', e)
-        raise
-
-
-def get_region_top_species(api_key: str, region_code: str) -> list[str]:
-    """Get top N species by observation count in region. Returns common names."""
-    obs = _fetch_region_observations(api_key, region_code)
-    if not obs:
-        return []
-    counter = Counter()
-    for o in obs:
-        com = (o.get('comName') or '').strip()
-        if com:
-            counter[com] += 1
-    return [name for name, _ in counter.most_common(TOP_N)]
-
-
-# Shared cache so Bird Directory regional scope and mapping suggestions do not duplicate HTTP.
-_REGION_TOP_CACHE: dict[tuple[str, str], tuple[float, list[str]]] = {}
-_REGION_TOP_TTL_SEC = 1800.0
-
-
-def get_region_top_species_cached(api_key: str, region_code: str) -> list[str]:
-    """Same as ``get_region_top_species`` but cached ~30 min per region and key suffix."""
-    now = time.monotonic()
-    key_suffix = api_key[-12:] if len(api_key) >= 12 else api_key
-    key = (region_code, key_suffix)
-    ent = _REGION_TOP_CACHE.get(key)
-    if ent is not None and (now - ent[0]) < _REGION_TOP_TTL_SEC:
-        return ent[1]
-    try:
-        top = get_region_top_species(api_key, region_code)
-    except Exception as e:
-        logger.warning('eBird regional top fetch failed: %s', e)
-        top = []
-    _REGION_TOP_CACHE[key] = (now, top)
-    return top
-
-
-def _ebird_to_birdlense(name: str) -> str:
-    """Map eBird common name to BirdLense canonical.
-
-    Checks ebird.species_mapping first (user overrides), then detection.species_mapping
-    (centralized name harmonization). This way Gray/Grey etc. work without duplicating
-    entries in both config sections.
-    """
-    ebird_map = app_config.get('ebird.species_mapping') or {}
-    if name in ebird_map:
-        return ebird_map[name]
-    detection_map = app_config.get('detection.species_mapping') or {}
-    return detection_map.get(name, name)
-
-
-def ebird_common_to_birdlense_name(name: str) -> str:
-    """Public alias for species matching (Bird Directory, filters)."""
-    return _ebird_to_birdlense(name)
+# Backward compat: tests clear ebird_region_service._REGION_TOP_CACHE
+# (same dict as ebird_region_core._REGION_TOP_CACHE)
 
 
 def get_region_comparison(user_species_names: list[str]) -> dict | None:
@@ -118,9 +34,9 @@ def get_region_comparison(user_species_names: list[str]) -> dict | None:
 
     region_code = _build_region_code()
     sorted_names = sorted(n for n in user_species_names if n)
-    names_hash = hashlib.sha256("|".join(sorted_names).encode()).hexdigest()[:32]
+    names_hash = hashlib.sha256('|'.join(sorted_names).encode()).hexdigest()[:32]
     key_suffix = api_key[-12:] if len(api_key) >= 12 else api_key
-    cache_key = f"ebird_region_comparison:{key_suffix}:{region_code}:{names_hash}"
+    cache_key = f'ebird_region_comparison:{key_suffix}:{region_code}:{names_hash}'
 
     found, cached_result = cache_get(cache_key)
     if found:
@@ -147,8 +63,7 @@ def get_region_comparison(user_species_names: list[str]) -> dict | None:
         cache_set(cache_key, result, ttl_seconds=14400)
         return result
 
-    # eBird names -> BirdLense canonical для сопоставления
-    region_mapped = [_ebird_to_birdlense(n) for n in region_top]
+    region_mapped = [ebird_common_to_birdlense_name(n) for n in region_top]
     region_set = {n.lower() for n in region_mapped}
     matched = [n for n in user_common if n.lower() in region_set]
 
@@ -158,7 +73,17 @@ def get_region_comparison(user_species_names: list[str]) -> dict | None:
         'regionTopCount': len(region_top),
         'matchCount': len(matched),
         'matchedSpecies': matched,
-        'regionTop': region_mapped,  # показываем BirdLense-имена
+        'regionTop': region_mapped,
     }
     cache_set(cache_key, result, ttl_seconds=14400)
     return result
+
+
+__all__ = [
+    '_REGION_TOP_CACHE',
+    '_build_region_code',
+    'ebird_common_to_birdlense_name',
+    'get_region_top_species',
+    'get_region_top_species_cached',
+    'get_region_comparison',
+]
