+++ app/web/services/ebird_util.py
@@ -1,13 +1,7 @@
 """Shared utilities for eBird-related services (export, region comparison)."""
 import re
 
-# Маппинг полных названий регионов на коды eBird
-REGION_NAME_TO_CODE = {
-    "moscow oblast": "MOS",
-    "moscow": "MO",
-    "московская область": "MOS",
-    "москва": "MO",
-}
+from ebird_region_core import REGION_NAME_TO_CODE
 
 
 def common_name_from_species(name: str) -> str:
