+++ docs/ROADMAP.md
@@ -228,7 +228,7 @@ Historical **simple → complex** checklist (all rows shipped). Cross-check [FEA
 
 **Execution lives in GitHub issues and the [BirdLense Hub — Roadmap](https://github.com/users/Gfermoto/projects/2) board**; this section is a pointer only.
 
-- Navigator: [#220](https://github.com/Gfermoto/BirdLense-Hub/issues/220). **Sub-issues** (board hierarchy): [#198](https://github.com/Gfermoto/BirdLense-Hub/issues/198) (**closed** — split `ui_routes` into domain `ui_*_routes`, see [ARCHITECTURE](./ARCHITECTURE.md)), [#201](https://github.com/Gfermoto/BirdLense-Hub/issues/201) (**closed** — processor modularization, HA `user_config` migration, PR [#237](https://github.com/Gfermoto/BirdLense-Hub/pull/237)), [#238](https://github.com/Gfermoto/BirdLense-Hub/issues/238) (processor **phase 2** — eBird import/DI, optional `RecordingSession`; MQTT queue persist + reconnect flush **done**), [#221](https://github.com/Gfermoto/BirdLense-Hub/issues/221) · [#222](https://github.com/Gfermoto/BirdLense-Hub/issues/222) · [#223](https://github.com/Gfermoto/BirdLense-Hub/issues/223) · [#224](https://github.com/Gfermoto/BirdLense-Hub/issues/224) (**closed** — MQTT outbound queue, `frame_processor` no sleep) · [#225](https://github.com/Gfermoto/BirdLense-Hub/issues/225). Re-link: `bash scripts/github-issue-link-subissues.sh 220 198 201 238 221 222 223 224 225`. RU: [ROADMAP.ru.md](./ROADMAP.ru.md) (wave D). Not this wave: [#164](https://github.com/Gfermoto/BirdLense-Hub/issues/164); Heimdall/HA backlog [#229](https://github.com/Gfermoto/BirdLense-Hub/issues/229)–[#234](https://github.com/Gfermoto/BirdLense-Hub/issues/234).
+- Navigator: [#220](https://github.com/Gfermoto/BirdLense-Hub/issues/220). **Sub-issues** (board hierarchy): [#198](https://github.com/Gfermoto/BirdLense-Hub/issues/198) (**closed** — split `ui_routes` into domain `ui_*_routes`, see [ARCHITECTURE](./ARCHITECTURE.md)), [#201](https://github.com/Gfermoto/BirdLense-Hub/issues/201) (**closed** — processor modularization, HA `user_config` migration, PR [#237](https://github.com/Gfermoto/BirdLense-Hub/pull/237)), [#238](https://github.com/Gfermoto/BirdLense-Hub/issues/238) (processor **phase 2** — опционально `RecordingSession`; MQTT queue + **ebird_region_core** **done**), [#221](https://github.com/Gfermoto/BirdLense-Hub/issues/221) · [#222](https://github.com/Gfermoto/BirdLense-Hub/issues/222) · [#223](https://github.com/Gfermoto/BirdLense-Hub/issues/223) · [#224](https://github.com/Gfermoto/BirdLense-Hub/issues/224) (**closed** — MQTT outbound queue, `frame_processor` no sleep) · [#225](https://github.com/Gfermoto/BirdLense-Hub/issues/225). Re-link: `bash scripts/github-issue-link-subissues.sh 220 198 201 238 221 222 223 224 225`. RU: [ROADMAP.ru.md](./ROADMAP.ru.md) (wave D). Not this wave: [#164](https://github.com/Gfermoto/BirdLense-Hub/issues/164); Heimdall/HA backlog [#229](https://github.com/Gfermoto/BirdLense-Hub/issues/229)–[#234](https://github.com/Gfermoto/BirdLense-Hub/issues/234).
 
 ---
