+++ docs/ROADMAP.ru.md
@@ -275,7 +275,7 @@ bash scripts/github-project-add-backlog-consilium.sh
 
 **Волна D (техдолг):** исполнение и чеклисты — **только в GitHub issues и на доске** [BirdLense Hub — Roadmap](https://github.com/users/Gfermoto/projects/2); здесь краткая отсылка.
 
-- Эпик-порядок: [#220](https://github.com/Gfermoto/BirdLense-Hub/issues/220). **Sub-issues** (колонка иерархии на доске): сводные [#198](https://github.com/Gfermoto/BirdLense-Hub/issues/198) (**закрыт** — нарезка `ui_routes` на доменные `ui_*_routes`, см. [ARCHITECTURE.ru.md](./ARCHITECTURE.ru.md)), [#201](https://github.com/Gfermoto/BirdLense-Hub/issues/201) (**закрыт** — декомпозиция процессора, миграция HA в `user_config`, PR [#237](https://github.com/Gfermoto/BirdLense-Hub/pull/237)), [#238](https://github.com/Gfermoto/BirdLense-Hub/issues/238) (processor **фаза 2** — eBird/DI, опционально `RecordingSession`; очередь MQTT при disconnect + слив после reconnect — **сделано**) и уровни [#221](https://github.com/Gfermoto/BirdLense-Hub/issues/221) · [#222](https://github.com/Gfermoto/BirdLense-Hub/issues/222) · [#223](https://github.com/Gfermoto/BirdLense-Hub/issues/223) · [#224](https://github.com/Gfermoto/BirdLense-Hub/issues/224) (**закрыт** — очередь MQTT + `frame_processor` без sleep) · [#225](https://github.com/Gfermoto/BirdLense-Hub/issues/225). Восстановить связи: `bash scripts/github-issue-link-subissues.sh 220 198 201 238 221 222 223 224 225`.
+- Эпик-порядок: [#220](https://github.com/Gfermoto/BirdLense-Hub/issues/220). **Sub-issues** (колонка иерархии на доске): сводные [#198](https://github.com/Gfermoto/BirdLense-Hub/issues/198) (**закрыт** — нарезка `ui_routes` на доменные `ui_*_routes`, см. [ARCHITECTURE.ru.md](./ARCHITECTURE.ru.md)), [#201](https://github.com/Gfermoto/BirdLense-Hub/issues/201) (**закрыт** — декомпозиция процессора, миграция HA в `user_config`, PR [#237](https://github.com/Gfermoto/BirdLense-Hub/pull/237)), [#238](https://github.com/Gfermoto/BirdLense-Hub/issues/238) (processor **фаза 2** — опционально `RecordingSession`; MQTT + **ebird_region_core** без `services.*` — **сделано**) и уровни [#221](https://github.com/Gfermoto/BirdLense-Hub/issues/221) · [#222](https://github.com/Gfermoto/BirdLense-Hub/issues/222) · [#223](https://github.com/Gfermoto/BirdLense-Hub/issues/223) · [#224](https://github.com/Gfermoto/BirdLense-Hub/issues/224) (**закрыт** — очередь MQTT + `frame_processor` без sleep) · [#225](https://github.com/Gfermoto/BirdLense-Hub/issues/225). Восстановить связи: `bash scripts/github-issue-link-subissues.sh 220 198 201 238 221 222 223 224 225`.
 - Вне волны D: [#164](https://github.com/Gfermoto/BirdLense-Hub/issues/164) (классификатор); Heimdall/HA — бэклог [#229](https://github.com/Gfermoto/BirdLense-Hub/issues/229)–[#234](https://github.com/Gfermoto/BirdLense-Hub/issues/234).
 
 **Критерий выхода из этапа "мелкие баги/хвосты":**
